Refactor fetch logic to module
