+++ app/services/cml.rb
@@ -1,6 +1,8 @@
 require 'open-uri'
 
 class CML
+  extend Base
+
   class << self
     CML_BASE_URL = 'https://crystalmathlabs.com/tracker/api.php'
 
@@ -84,32 +86,6 @@ class CML
         'smithing_xp'    => xps[14],
         'mining_xp'      => xps[15],
         'runecraft_xp'   => xps[21] }
-    end
-
-    private
-
-    def fetch(uri, max_attempts = 3)
-      raise TypeError, 'uri not an URI object' unless URI == uri
-
-      openuri_params = {
-        # Timeout durations for HTTP connection.
-        # 5 seconds should be max for opening and reading a connection.
-        open_timeout: 5,
-        read_timeout: 5
-      }
-
-      attempt = 0
-
-      begin
-        attempt += 1
-        return uri.read(openuri_params)
-      rescue OpenURI::HTTPError => e
-        # 404, no content.
-      rescue SocketError, Net::ReadTimeout => e
-        Rails.logger.warn "CML cannot be reached: #{e}"
-        sleep 2
-        retry if attempt < max_attempts
-      end
     end
   end
 end
